# src/server/openapi.yaml
openapi: 3.0.3
info:
  title: وسِيطي API
  description: Real Estate CRM & Platform Integration API
  version: 1.0.0
  contact:
    name: Waseety API Support
    email: api@waseety.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.waseety.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [agent, manager, admin, owner]
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    Workspace:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        owner_id:
          type: string
          format: uuid
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: تسجيل مستخدم جديد
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - workspace_name
              properties:
                name:
                  type: string
                  example: "أحمد محمد"
                email:
                  type: string
                  format: email
                  example: "ahmed@example.com"
                phone:
                  type: string
                  example: "0501234567"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123!"
                workspace_name:
                  type: string
                  example: "مكتب أحمد العقاري"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          workspace:
                            $ref: '#/components/schemas/Workspace'
                          tokens:
                            type: object
                            properties:
                              access_token:
                                type: string
                              refresh_token:
                                type: string
                              expires_in:
                                type: integer
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: تسجيل الدخول
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                workspace_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user:
                            $ref: '#/components/schemas/User'
                          tokens:
                            type: object
                            properties:
                              access_token:
                                type: string
                              refresh_token:
                                type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...